common: not copy constants.go for compatibility
